feat: update routing and implement login page; refactor logout behavior and enhance cart functionality
